Finalizing Shop
I have made the finishing touches to the shop so that it rerolls a new set of abilities for the player to buy after they have bought all of the currently available items. I also fixed up the randomness and plan to continue working on the abilities the player can choose from.